Merge branch 'master' of https://github.com/Bobay-Ochman/ConSpeciFix

+++ web/write_core.py
@@ -17,7 +17,7 @@ tmp={}
 orthologFD=open(PATH_TO_UPLOAD + 'orthologs.txt',"r")
 core = {}
 for line in orthologFD:
-	l = line.remove('\n').split('\t')
+	l = line.strip('\n').split('\t')
 	orthoName = l[0]
 	l.pop(0)
 	core[orthoName] = l[:] #deep copy l
@@ -69,12 +69,18 @@ for sp in species:
 		pass
 	nb=0
 	for lili in core:
-		#### Change to only do this if lili contains 'gene####'
-		g=open(PATH_TO_UPLOAD+ 'align/' + lili + ".fa","w")
-		for id in core[lili]:
-			g.write(">" + id + "\n" + seq[sp][id]  )
-		g.truncate()
-		g.close()	
+		# Only write the core gene if 'gene###' is in the list
+		flag = False
+		for geneId in core[lili]:
+			if 'gene' in geneId:
+				falg=True
+		if flag:
+			nb = nb+1
+			g=open(PATH_TO_UPLOAD+ 'align/' + lili + ".fa","w")
+			for id in core[lili]:
+				g.write(">" + id + "\n" + seq[sp][id]  )
+			g.truncate()
+			g.close()	
 
 	critInfoFD = open(PATH_TO_UPLOAD+'crit_stats.txt','a')
 	critInfoFD.write('Number of core genes orthologous to your genome: '+str(nb)+'\n')
